requirements.txt
Django==5.2.8
Pillow==12.0.0
black==25.12.0
celery-redbeat==2.3.3
celery[redis]==5.6.1
crispy-tailwind==1.0.3
django-allauth[socialaccount]==65.13.1
django-crispy-forms==2.5
django-debug-toolbar==6.1.0
django-environ==0.12.0
django-extensions==4.1
django-post_office==3.11
django-ses==4.5.0
django-stubs[compatible-mypy]==5.2.8
djangoql==0.18.1
gunicorn==23.0.0
psycopg[binary]==3.3.2
pyjwt[crypto]==2.10.1
requests==2.32.5
whitenoise==6.11.0
django-structlog[commands,celery]==10.0.0
django-recaptcha==4.1.0
# API
djangorestframework==3.16.1
drf-spectacular==0.29.0
# Authentication & Security
django-otp==1.6.3
qrcode==8.2
django-fernet-encrypted-fields
# Monitoring
httpx==0.28.1
dnspython==2.8.0
# Payments
django-paddle-billing==0.1.14
# Utilities
python-slugify==8.0.4